+++ src/AgentAcademy.Server/Program.cs
@@ -247,6 +247,7 @@ builder.Services.AddSingleton<ICommandHandler, CloseRoomHandler>();
 builder.Services.AddSingleton<ICommandHandler, CreateRoomHandler>();
 builder.Services.AddSingleton<ICommandHandler, ReopenRoomHandler>();
 builder.Services.AddSingleton<ICommandHandler, InviteToRoomHandler>();
+builder.Services.AddSingleton<ICommandHandler, ReturnToMainHandler>();
 builder.Services.AddSingleton<ICommandHandler, MergeTaskHandler>();
 builder.Services.AddSingleton<ICommandHandler, CancelTaskHandler>();
 builder.Services.AddSingleton<ICommandHandler, ShellCommandHandler>();